dərs 32 çalışmaları

+++ spring web MVC 2/spring-jpa-ayaan/src/main/java/az/developia/springjpaayaan/BookController.java
@@ -1,5 +1,7 @@
 package az.developia.springjpaayaan;
 
+import java.util.List;
+
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 @RequestMapping(path = "/kitab")
@@ -31,8 +34,13 @@ public class BookController {
 	}
 
 	@GetMapping(path = "/{id}")
-	public String delete(@PathVariable Integer id) {
-		bookRepository.deleteById(id);
+	public String delete(@PathVariable Integer id, Model model2) {
+		if (bookRepository.findById(id).isPresent()) {
+			bookRepository.deleteById(id);
+		}else {
+			model2.addAttribute("message", "bu id-li kitab tapilmadi, id = "+id);
+			return "errorumuz";
+		}
 
 		return "redirect:/kitab/list";
 	}
@@ -55,5 +63,19 @@ public class BookController {
 		bookRepository.save(book);
 
 		return "redirect:/kitab/list";
+	}
+
+	@GetMapping(path = "/search") // http://localhost:9090/search?s=(istenilen sey)
+	public String search(@RequestParam(name = "s") String columnName, Model model2) {
+
+		// List<Book> books = bookRepository.findAllByNameContainingOrPriceContaining(columnName,columnName);
+		// List<Book> books = bookRepository.hamisiniTapinLimit(2, 2); // 2-ci məlumatdan sonra 2 dənə olan məlumatı göstərir.
+		bookRepository.deleteByName(columnName);
+		List<Book> books = bookRepository.findAll();
+
+		model2.addAttribute("kitab", books);
+
+		return "books";
+
 	}
 }
